Fixing hero layout on responsive devices
Give this a test and see how it goes. Usually per device its better to use unique dimensions per, but in this case it might be good with just this (will do per device if we end up needing it). Will need to provide testing with minting as well (seeing how the mint slider functions and all if its gonna be placed in the hero block - this wont affect the minting process at all)